+++ src/middlewares/multerMiddleware.js
@@ -1,0 +1,18 @@
+
+const path = require("path");
+const multer = require("multer");
+
+// Multer - manejo del almacenamiento
+const storage = multer.diskStorage({
+	destination: (req , file , cb) => {
+		cb (null , path.resolve(__dirname , "../../public/images/productos"));
+	},
+	filename: (req , file , cb) => {
+		cb (null , file.fieldname + "-" + Date.now() + path.extname(file.originalname))
+	}
+});
+
+// Instanciar multer para manejar los métodos
+const upload = multer ({ storage });
+
+module.exports = upload;
